+++ docs/outputs.md
@@ -39,12 +39,12 @@ Tuning IO parameters can be passed to Parthenon through the use of environment v
 |  Environment Variable | Default | Value | Description |
 |---|---|---|---|
 | H5_sieve_buf_size | disabled | int | Sets the maximum size of the data sieve buffer, in bytes. The value should be equal to a multiple of the disk block size. If no value is set then the default is 256 KiB. |
-| H5_meta_block_size | enabled | int | Sets the minimum metadata block size, in bytes. If no value is   set then the default is 16 MiB. |
-| H5_alignment_threshold | disabled | int | The threshold value, in bytes, of H5Pset_alignment. Setting to 0 forces  everything to be aligned. If a value is not set then the default is 1 MiB. Setting the environment variable automatically enables alignment. |
-| H5_alignment_alignment | disabled | int | The alignment value, in bytes, of H5Pset_alignment. If a value is not set then the default is 16 MiB.  Setting the environment variable automatically enables alignment.  H5Pset_alignment sets the alignment properties of a file access property list. Choose an alignment  that is a multiple of the disk block size, enabling this usually shows better performance on GPFS.   Enabling may increase the file size significantly. |
-| H5_defer_metadata_flush | disabled | int | Value of 1 enables deferring metadata flush. Value of 0 disables. |
+| H5_meta_block_size | disabled | int | Sets the minimum metadata block size, in bytes. If no value is set then the default is 8 MiB. May help performance if enabled. |
+| H5_alignment_threshold | disabled | int | The threshold value, in bytes, of H5Pset_alignment. Setting to 0 forces everything to be aligned. If a value is not set then the default is 0. Setting the environment variable automatically enables alignment. |
+| H5_alignment_alignment | disabled | int | The alignment value, in bytes, of H5Pset_alignment. If a value is not set then the default is 8 MiB.  Setting the environment variable automatically enables alignment.  H5Pset_alignment sets the alignment properties of a file access property list. Choose an alignment that is a multiple of the disk block size, enabling this usually shows better performance on parallel file systems. However, enabling may increase the file size significantly. |
+| H5_defer_metadata_flush | disabled | int | Value of 1 enables deferring metadata flush. Value of 0 disables. Experiment with before using. |
 | MPI_access_style | enabled | string | Specifies the manner in which the file will be accessed until the file is closed. Default is "write_once" |
-| MPI_collective_buffering | disabled | int | Value of 1 enables MPI collective buffering. Value of 0 disables. |
+| MPI_collective_buffering | disabled | int | Value of 1 enables MPI collective buffering. Value of 0 disables. Experiment with before using. |
 | MPI_cb_block_size | N/A | int | Sets the block size, in bytes, to be used for collective buffering file access. Default is 1 MiB. |
 | MPI_cb_buffer_size | N/A | int | Sets the total buffer space, in bytes, that can be used for collective buffering on each target node,  usually a multiple of cb_block_size. Default is 4 MiB. |
